telegram_bot.py
# telegram_bot.py

import os
import telegram
from telegram.ext import Dispatcher, CommandHandler, MessageHandler, Filters

TELEGRAM_TOKEN = os.getenv('TELEGRAM_TOKEN')
if not TELEGRAM_TOKEN:
    raise ValueError("TELEGRAM_TOKEN environment variable is not set")

bot = telegram.Bot(token=TELEGRAM_TOKEN)
dispatcher = Dispatcher(bot, None, workers=0)

def start(update, context):
    context.bot.send_message(chat_id=update.effective_chat.id, text="Welcome to the Crypto News and DCA Bot!")

def help_command(update, context):
    context.bot.send_message(chat_id=update.effective_chat.id, text="Available commands:\n/start\n/help\n/getnews\n/setdca\n/status")

def get_news(update, context):
    # Implement get_news functionality here
    pass

def setdca(update, context):
    # Implement setdca functionality here
    pass

def confirmdca(update, context):
    # Implement confirmdca functionality here
    pass

def status(update, context):
    # Implement status functionality here
    pass

# Add handlers to dispatcher
dispatcher.add_handler(CommandHandler('start', start))
dispatcher.add_handler(CommandHandler('help', help_command))
dispatcher.add_handler(CommandHandler('getnews', get_news))
dispatcher.add_handler(CommandHandler('setdca', setdca))
dispatcher.add_handler(CommandHandler('confirmdca', confirmdca))
dispatcher.add_handler(CommandHandler('status', status))

def setup_webhook(app):
    @app.route('/webhook', methods=['POST'])
    def webhook():
        update = telegram.Update.de_json(request.get_json(force=True), bot)
        dispatcher.process_update(update)
        return 'OK'

# This function can be called to start the bot
def run_bot():
    # Implement polling or webhook setup here
    pass