requirements.txt
Flask
python-telegram-bot==13.15
requests
beautifulsoup4
schedule
openai
gunicorn
python-dotenv
playwright
pydub